Dashboard/requirement.txt
bcrypt             3.1.4
blinker            1.4
certifi            2021.10.8
cffi               1.15.0
charset-normalizer 2.0.8
click              8.0.3
Flask              1.0
Flask-Bcrypt       0.7.1
Flask-Login        0.4.1
Flask-Mail         0.9.1
Flask-SQLAlchemy   2.3.2
Flask-WTF          0.14.2
greenlet           1.1.2
idna               3.3
importlib-metadata 4.8.2
itsdangerous       0.24
Jinja2             2.10
lxml               4.6.4
MarkupSafe         2.0.1
numpy              1.21.4
pandas             1.3.4
pandas-datareader  0.10.0
Pillow             5.3.0
pip                21.2.2
pycparser          2.18
python-dateutil    2.8.2
pytz               2021.3
requests           2.26.0
setuptools         58.0.4
six                1.11.0
SQLAlchemy         1.2.7
typing_extensions  4.0.0
urllib3            1.26.7
Werkzeug           0.14.1
wheel              0.37.0
WTForms            2.1
zipp               3.6.0
